correctifs mentorat

+++ backend/controllers/book.js
@@ -19,18 +19,27 @@ exports.createBook = async (req, res) => {
 };
 
 exports.modifyBook = async (req, res) => {
-    const newImg = await compressImg(req.file.buffer);
+    delete req.body.ratings;
+    if(req.file) {
+        req.body.imageUrl = await compressImg(req.file.buffer)
+    }
     Book.findOne({ _id: req.params.id })
         .then((book) => {
             if(book.userId != req.auth.userId) {
                 res.status(401).json({ message: 'Non autorisé !' });
             } else {
-                const filename = book.imageUrl;
-                fs.unlink(`images/${filename}`, () => {                 
-                    Book.updateOne({ _id: req.params.id }, { ...req.body, imageUrl: newImg, _id: req.params.id })
-                        .then(() => res.status(200).json({ message: 'Livre modifié !'}))
-                        .catch(error => res.status(400).json({ error }));
-                })
+                if(req.file) {
+                    const filename = book.imageUrl;
+                    fs.unlink(`images/${filename}`, () => {                 
+                        Book.updateOne({ _id: req.params.id }, { ...req.body })
+                            .then(() => res.status(200).json({ message: 'Livre modifié !'}))
+                            .catch(error => res.status(400).json({ error }));
+                    })
+                } else {
+                    Book.updateOne({ _id: req.params.id }, { ...req.body })
+                            .then(() => res.status(200).json({ message: 'Livre modifié !'}))
+                            .catch(error => res.status(400).json({ error }));
+                }
             }
         })
         .catch((error) => {
